Further work on tests, added command line parsing for local store directory.
