+++ .vscode/launch.json
@@ -10,7 +10,6 @@
             "request": "launch",
             "program": "${workspaceRoot}/app/server/main.js",
             "stopOnEntry": false,
-            "args": [],
             "runtimeExecutable": "${workspaceRoot}/node_modules/.bin/electron",
             "windows": {
                 "runtimeExecutable": "${workspaceRoot}/node_modules/.bin/electron.cmd"
@@ -22,6 +21,19 @@
                 "NODE_ENV": "development"
             },
             "sourceMaps": true,
+            "protocol": "inspector"
+        },
+        {
+            "name": "Debug integration tests",
+            "type": "node",
+            "request": "launch",
+            "program": "${workspaceRoot}/node_modules/mocha/bin/_mocha",
+            "runtimeExecutable": "node",
+            "stopOnEntry": false,
+            "args": ["-t", "20000", "--compilers", "ts:ts-node/register", "${workspaceRoot}/tests/integration/integration.ts"],
+            "env": { "NODE_ENV": "test"},
+            "cwd": "${workspaceRoot}",
+            "sourceMaps": false,
             "protocol": "inspector"
         }
     ]
